Scope draw-order fix + exact motion-profile conformance for both axes
The scope's knife-tip trace vanished during every cut window: tip and
material velocities are EQUAL in sync, and the material line was drawn
second, painting pixel-perfectly over the tip trace at the exact moment
the scope exists to show. Material now draws first and wide, tip thin
on top - the sync lock reads as red riding centered on blue. Scope time
axis moved to Float64 (Float32 aliases the 1 ms sampling after ~2.5 h
of machine time - endurance runs are this rig's job).

New tests/motion-conformance.test.ts proves every motor follows its
motion profile, not just its limits:
- feed axis 0->cruise ramp compared point-by-point against the analytic
  MotionProfileSolver evalAt profile (both accel shapes)
- graceful stop lands within 0.01 mm of the park target; braked
  distance matches commitStopDist up to commit-scan quantization
- knife angular velocity matches the cam gearing law dtheta/dphi * v on
  every scan of full runs (start ramp, steady, stop) for 4 cam types,
  and never rotates backwards

Making those tests pass exposed and fixed three real tracker flaws:
- jerk flips quantized to whole scans drifted the ramp ~0.7% off the
  analytic S-curve; the crossing scan now applies a FRACTIONAL jerk
  step landing the settle velocity exactly on the setpoint (trapezoidal
  integration provably conserves it from there)
- position integrated with the rectangle rule, under-travelling every
  brake ramp by v*dt/2 (millimeters at speed); now trapezoidal,
  consistent with the velocity integration - stops land dead on target
- at the stop landing the pinned master made the dead-beat follower
  command a tiny reverse; a rotary knife axis is one-directional, the
  drive now clamps omega >= 0

102/102 tests green.

+++ src/engine/SimulationEngine.ts
@@ -85,9 +85,12 @@ export class SimulationEngine {
   public scopeCut = false;
 
   // --- Data Block: Scope ring buffer (1 kHz machine-time recording) ---
+  // The time axis is Float64: with Float32 the 1 ms sample spacing falls
+  // below the float resolution after ~2.5 h of machine time and the
+  // scope time base aliases — endurance runs are exactly this rig's job.
   public scope = {
     cap: SCOPE_CAP, n: 0, i: 0,
-    t: new Float32Array(SCOPE_CAP),
+    t: new Float64Array(SCOPE_CAP),
     vMat: new Float32Array(SCOPE_CAP),
     vTip: new Float32Array(SCOPE_CAP),
     errDeg: new Float32Array(SCOPE_CAP),
